pyproject.toml
[project]
name = "jirascope"
version = "0.1.0"
description = "AI-powered Jira work item analysis and management tool"
authors = [
    {name = "Alex Shumilov",email = "alex.shumilov@rationalexponent.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "qdrant-client>=1.7.0,<2.0.0",
    "pydantic>=2.5.0,<3.0.0",
    "click>=8.1.0,<9.0.0",
    "rich>=13.7.0,<14.0.0",
    "fastapi>=0.104.0,<0.105.0",
    "uvicorn>=0.24.0,<0.25.0",
    "httpx>=0.25.0,<0.26.0",
    "pyyaml>=6.0,<7.0",
    "pandas>=2.1.0,<3.0.0",
    "numpy>=1.24.0,<2.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "anthropic>=0.8.0,<0.9.0"
]

[project.scripts]
jirascope = "cli.main:cli"

[tool.poetry]
packages = [{include = "jirascope", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"